+++ Viz.py
@@ -1,6 +1,7 @@
-from PlaneModel import *
+from plane import PlaneModel
 from mesa.visualization.modules import CanvasGrid
 from mesa.visualization.ModularVisualization import ModularServer
+import methods
 
 def passenger_portrayal(agent):
     portrayal = {"Shape": "circle",
@@ -14,6 +15,6 @@ grid = CanvasGrid(passenger_portrayal, 21, 7, 840, 310)
 server = ModularServer(PlaneModel,
                        [grid],
                        "Boarding Simulation",
-                       {random})
+                       {"method": methods.random})
 server.port = 8521 # The default
 server.launch()
