+++ queue.py
@@ -1,0 +1,10 @@
+from mesa.time import BaseScheduler
+
+
+class QueueActivation(BaseScheduler):
+    def step(self):
+        for agent in self.agent_buffer():
+            agent.step()
+        self.time += 1
+        self.steps += 1
+
